requirements.txt
numpy
pandas
scanpy
umap-learn
seaborn
matplotlib
scipy
networkx
plotnine
pynndescent
leidenalg
louvain
igraph